requirements.txt
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==7.0.5
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
click-default-group==1.2.4
cmdop==0.1.43
distro==1.9.0
fastapi==0.135.1
gitdb==4.0.12
GitPython==3.1.46
grpcio==1.78.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jiter==0.13.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
linkify-it-py==2.1.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdit-py-plugins==0.5.0
mdurl==0.1.2
narwhals==2.17.0
numpy==2.4.3
openai==2.26.0
packaging==26.0
pandas==2.3.3
pillow==12.1.1
platformdirs==4.9.4
pluggy==1.6.0
protobuf==6.33.5
pyarrow==23.0.1
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pyte==0.8.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-telegram-bot==22.6
pytz==2026.1.post1
referencing==0.37.0
requests==2.32.5
rich==14.3.3
rpds-py==0.30.0
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
soupsieve==2.8.3
sqlite-fts4==1.0.3
sqlite-utils==3.39
starlette==0.52.1
streamlit==1.55.0
tabulate==0.10.0
tenacity==9.1.4
textual==8.1.1
toml==0.10.2
tornado==6.5.4
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uc-micro-py==2.0.0
urllib3==2.6.3
uvicorn==0.41.0
watchdog==6.0.0
wcwidth==0.6.0
